Added prompt endpoints, simple auth and prompt tracking.

+++ Domain/Dto/Llm/LlmPriceDto.cs
@@ -2,5 +2,5 @@
 
 public record LlmPriceDto(
     Guid Id,
-    long MillionInputTokenPrice,
-    long MillionOutputTokenPrice);
+    int MillionInputTokenPrice,
+    int MillionOutputTokenPrice);
